+++ crates/ruma-events/src/key/verification/ready.rs
@@ -12,7 +12,7 @@ use super::{Relation, VerificationMethod};
 #[derive(Clone, Debug, Deserialize, Serialize, EventContent)]
 #[cfg_attr(not(feature = "unstable-exhaustive-types"), non_exhaustive)]
 #[ruma_event(type = "m.key.verification.ready", kind = ToDevice)]
-pub struct ReadyToDeviceEventContent {
+pub struct ToDeviceReadyEventContent {
     /// The device ID which is initiating the request.
     pub from_device: DeviceIdBox,
 
@@ -27,8 +27,8 @@ pub struct ReadyToDeviceEventContent {
     pub transaction_id: String,
 }
 
-impl ReadyToDeviceEventContent {
-    /// Creates a new `ReadyToDeviceEventContent` with the given device ID, verification methods and
+impl ToDeviceReadyEventContent {
+    /// Creates a new `ToDeviceReadyEventContent` with the given device ID, verification methods and
     /// transaction ID.
     pub fn new(
         from_device: DeviceIdBox,
@@ -75,7 +75,7 @@ mod tests {
     use ruma_identifiers::{event_id, DeviceIdBox};
     use serde_json::{from_value as from_json_value, json, to_value as to_json_value};
 
-    use super::{ReadyEventContent, ReadyToDeviceEventContent};
+    use super::{ReadyEventContent, ToDeviceReadyEventContent};
     use crate::key::verification::{Relation, VerificationMethod};
 
     #[test]
@@ -106,7 +106,7 @@ mod tests {
             "transaction_id": "456",
         });
 
-        let content = ReadyToDeviceEventContent {
+        let content = ToDeviceReadyEventContent {
             from_device: device,
             transaction_id: "456".to_owned(),
             methods: vec![VerificationMethod::SasV1],
@@ -149,8 +149,8 @@ mod tests {
         });
 
         assert_matches!(
-            from_json_value::<ReadyToDeviceEventContent>(json_data).unwrap(),
-            ReadyToDeviceEventContent {
+            from_json_value::<ToDeviceReadyEventContent>(json_data).unwrap(),
+            ToDeviceReadyEventContent {
                 from_device,
                 transaction_id,
                 methods,